requirements.txt
pandas
mysql.connector
os
tkinter
sqlite3